previous_works/deepfake-detector-main/requirements.txt
streamlit
tensorflow-cpu
numpy
Pillow
huggingface_hub
